pytorch warning fix
home/aishutin/.local/lib/python3.6/site-packages/torch/optim/lr_scheduler.py:82: UserWarning: Detected call of `lr_scheduler.step()` before `optimizer.step()`. In PyTorch 1.1.0 and later, you should call them in the opposite order: `optimizer.step()` before `lr_scheduler.step()`.  Failure to do this will result in PyTorch skipping the first value of the learning rate schedule.See more details at https://pytorch.org/docs/stable/optim.html#how-to-adjust-learning-rate

+++ bert/train.py
@@ -44,7 +44,7 @@ def config():
     batch_size = 16  # 16, 32 are recommended in the paper
 
     model_name = 'BERTLike'
-    model_config = {'bert_type': 'bert-base-uncased', 'lang': 'en', 'float_mode': 'fp16'}
+    model_config = {'bert_type': 'bert-base-uncased', 'lang': 'en', 'float_mode': 'fp32'}
 
     dataset_names = ['en-twitt-corpus' if model_config['lang'] == 'en' else 'ru-opendialog-corpus']
     max_dataset_size = int(1e5)
